+++ ferritin-amplify/src/amplify/mod.rs
@@ -8,5 +8,6 @@
 //!
 pub mod amplify;
 pub mod config;
+pub mod encoder;
 pub mod rotary;
 // pub mod loss;
